Share release metadata with the app
