+++ apps/app/src/components/settings/changelog-preview.ts
@@ -1,4 +1,5 @@
 import changelogSource from "../../../../../CHANGELOG.md?raw";
+export { RELEASE_META } from "../../../../../changelog-metadata";
 
 const LATEST_CHANGELOG_SOURCE_URL =
   "https://raw.githubusercontent.com/get-bb/bb/main/CHANGELOG.md";
@@ -17,48 +18,6 @@ interface ChangelogEntry {
   lede: ChangelogBlock[];
   sections: ChangelogSection[];
 }
-
-interface ChangelogReleaseMeta {
-  date: string;
-  headline: string;
-}
-
-export const CHANGELOG_RELEASE_META: Record<string, ChangelogReleaseMeta> = {
-  "0.39.0": {
-    date: "August 19, 2026",
-    headline: "Faster large threads and a long list of fixes",
-  },
-  "0.38.0": {
-    date: "August 15, 2026",
-    headline: "Extensions Page and Plugin Marketplaces",
-  },
-  "0.37.0": {
-    date: "August 11, 2026",
-    headline: "A much faster mobile app",
-  },
-  "0.36.0": {
-    date: "August 8, 2026",
-    headline: "Fixes and improvements",
-  },
-  "0.35.0": { date: "August 4, 2026", headline: "Plugins" },
-  "0.34.0": {
-    date: "July 28, 2026",
-    headline: "Fresher models, cross-provider questions",
-  },
-  "0.33.0": {
-    date: "July 21, 2026",
-    headline: "Quieter updates and safer approvals",
-  },
-  "0.0.31": { date: "July 17, 2026", headline: "Splits for everyone" },
-  "0.0.30": {
-    date: "July 14, 2026",
-    headline: "Multi-machine workflows and bb Connect",
-  },
-  "0.0.29": {
-    date: "July 9, 2026",
-    headline: "More agents, more models, redesigned Settings",
-  },
-};
 
 export function parseChangelogEntries(source: string): ChangelogEntry[] {
   const entries: ChangelogEntry[] = [];
