+++ changelog-metadata.ts
@@ -1,0 +1,59 @@
+type ReleaseMeta = {
+  date: string;
+  headline: string;
+};
+
+export const RELEASE_META: Record<string, ReleaseMeta> = {
+  "0.42.0": {
+    date: "September 5, 2026",
+    headline: "Account Pooler, push notifications, and a new plugin catalog",
+  },
+  "0.41.0": {
+    date: "September 1, 2026",
+    headline: "Scheduled sends, concurrency limits, and a rebuilt mobile app",
+  },
+  "0.40.0": {
+    date: "August 26, 2026",
+    headline: "File Editor, quick palette, and agent providers",
+  },
+  "0.39.0": {
+    date: "August 19, 2026",
+    headline: "Faster large threads and a long list of fixes",
+  },
+  "0.38.0": {
+    date: "August 15, 2026",
+    headline: "Extensions Page and Plugin Marketplaces",
+  },
+  "0.37.0": {
+    date: "August 11, 2026",
+    headline: "A much faster mobile app",
+  },
+  "0.36.0": {
+    date: "August 8, 2026",
+    headline: "Fixes and improvements",
+  },
+  "0.35.0": {
+    date: "August 4, 2026",
+    headline: "Plugins",
+  },
+  "0.34.0": {
+    date: "July 28, 2026",
+    headline: "Fresher models, cross-provider questions",
+  },
+  "0.33.0": {
+    date: "July 21, 2026",
+    headline: "Quieter updates and safer approvals",
+  },
+  "0.0.31": {
+    date: "July 17, 2026",
+    headline: "Splits for everyone",
+  },
+  "0.0.30": {
+    date: "July 14, 2026",
+    headline: "Multi-machine workflows and bb Connect",
+  },
+  "0.0.29": {
+    date: "July 9, 2026",
+    headline: "More agents, more models, redesigned Settings",
+  },
+};
